requirements.txt
asteval==0.9.32
audioread==3.0.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
expelliarmus==1.1.12
fonttools==4.50.0
h5py==3.10.0
idna==3.6
importRosbag==1.0.4
joblib==1.3.2
kiwisolver==1.4.5
lava-nc==0.9.0
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.42.0
matplotlib==3.8.3
msgpack==1.0.8
networkx==2.8.7
numba==0.59.1
numpy==1.26.4
packaging==24.0
pbr==6.0.0
pillow==10.2.0
platformdirs==4.2.0
pooch==1.8.1
pycparser==2.21
pyparsing==3.1.2
python-dateutil==2.9.0.post0
requests==2.31.0
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
soundfile==0.12.1
soxr==0.3.7
threadpoolctl==3.3.0
tonic==1.4.3
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
